Updates to the applet and test scripts
I have updated and tested the card and jacket applets as per the SOW.

Note: you will have to re-build each applet.

I have included the java source code of each of the test scripts to act
as a guide when testing across there.

You will find a method initialise() in both the CardTest.java and
JacketTest.java files which implements
the install and personalisation of the applets. This can be used as a
guide.

All the test scripts run using our test environment.

I have made some small correction to the applet specification as well.